reorganized file structure
Also, added new functions into readme.
