+++ jest.config.ts
@@ -126,7 +126,7 @@ export default {
 
 	// A list of paths to directories that Jest should use to search for files in
 	roots: [
-		"src",
+		"test",
 	],
 
 	// Allows you to use a custom runner instead of Jest's default test runner
