+++ src/comparisons/disjoint.function.ts
