+++ src/comparisons/equivalence.function.ts
