+++ src/operations/difference.function.ts
