+++ src/operations/intersection.function.ts
