+++ src/operations/xor.function.ts
