+++ test/operations/union.function.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
-import { equivalence, union } from '../index';
-import { empty, setA, setB, setC } from './constants/testing-constants';
+import { equivalence, union } from '../../src';
+import { empty, setA, setB, setC } from '../constants/testing-constants';
 
 describe('union', () => {
 	const unionAB = new Set<number>([ 0, 1, 2, 3, 4, 5 ]);
