+++ tsconfig.json
@@ -30,7 +30,6 @@
 	],
 
 	"exclude": [
-		"node_modules",
-		"src/test"
+		"node_modules"
 	]
 }
